fix: keep generated imports out of modules `$app/server` reaches
