+++ packages/kit/test/mocks/generated-server.js
@@ -1,7 +1,7 @@
 // Stub for `<sveltekit:generated>/server.js`, which `sync` writes from the user's config
 
-/** @type {import('types').SSROptions} */
-export const options = /** @type {any} */ ({});
+/** @type {any} */
+export const options = {};
 
 export function get_hooks() {
 	return Promise.resolve({});
